% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-avg.R
\name{avg}
\alias{avg}
\title{Average}
\description{
The \code{avg} function returns the average of a column
within each group.
}
\section{Function Name}{

\itemize{
\item \code{avg}
}
}

\section{Arguments}{

\itemize{
\item The name of a numerical column to average (required).
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)

dbGetPlot(con, "
  visualize
    origin as x,
    avg(horsepower) as y
  from cars
  group by
    origin
  using bars
")

}
