% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-bars.R
\name{bars}
\alias{bars}
\title{Bars}
\description{
Documents the aliases, aesthetics, and qualifiers for the bar geom.
}
\section{Aliases}{

\itemize{
\item \code{bar}
\item \code{bars}
}
}

\section{Aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
}
}

\section{Qualifiers}{

\itemize{
\item \code{horizontal}: orients the bars horizontally.
\item \code{unstacked}: doesn't stack overlapping bars.
\item \code{vertical}: orients the bars vertically.
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    bin(miles_per_gallon) as x,
    count(*) as y
  from cars
  group by
    bin(miles_per_gallon)
  using bars
")

dbGetPlot(con, "
  visualize
    bin(miles_per_gallon) as y,
    count(*) as x
  from cars
  group by
    bin(miles_per_gallon)
  using horizontal bars
")

}
