% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-bin.R
\name{bin}
\alias{bin}
\title{Bin}
\description{
The \code{bin} function creates equal-width bins for a
column; original values are transformed into bin-center values.
}
\section{Function Name}{

\itemize{
\item \code{bin}
}
}

\section{Arguments}{

\itemize{
\item The name of a numerical column to bin (required).
\item The number of bins (optional, default \code{30}).
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)

dbGetPlot(con, "
  visualize
    bin(miles_per_gallon) as x,
    count(*) as y
  from cars
  group by
    bin(miles_per_gallon)
  using bars
")

dbGetPlot(con, "
  visualize
    bin(miles_per_gallon, 10) as x,
    count(*) as y
  from cars
  group by
    bin(miles_per_gallon, 10)
  using bars
")

}
