% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-boxes.R
\name{boxes}
\alias{boxes}
\title{Boxes}
\description{
Documents the aliases, aesthetics, and qualifiers for the box geom.
}
\section{Aliases}{

\itemize{
\item \code{box}
\item \code{boxes}
}
}

\section{Aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
}
}

\section{Qualifiers}{

\itemize{
\item \code{horizontal}: orients the boxes horizontally.
\item \code{vertical}: orients the boxes vertically.
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    origin as x,
    miles_per_gallon as y
  from cars
  using boxes
")

dbGetPlot(con, "
  visualize
    origin as y,
    miles_per_gallon as x
  from cars
  using horizontal boxes
")

}
