% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-count.R
\name{count}
\alias{count}
\title{Count}
\description{
The \code{count} function returns the number of rows in each
group.
}
\section{Function Name}{

\itemize{
\item \code{count}
}
}

\section{Arguments}{

\itemize{
\item \code{*} (required).
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)

dbGetPlot(con, "
  visualize
    origin as x,
    count(*) as y
  from cars
  group by
    origin
  using bars
")

}
