% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-linear.R
\name{linear}
\alias{linear}
\title{Linear}
\description{
The \code{linear} scale linearly maps data values to a visual property.
It is the default scale for numerical aesthetic mappings.
}
\section{Function Name}{

\itemize{
\item \code{linear}
}
}

\section{Arguments}{

\itemize{
\item The name of an aesthetic to scale (required).
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)

# explicit linear scales
dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using points
  scale by
    linear(x),
    linear(y)
")

# default linear scales
dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using points
")

}
