% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-lines.R
\name{lines}
\alias{lines}
\title{Lines}
\description{
Documents the aliases, aesthetics, and qualifiers for the line geom.
}
\section{Aliases}{

\itemize{
\item \code{line}
\item \code{lines}
}
}

\section{Aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
}
}

\section{Qualifiers}{

\itemize{
\item \code{horizontal}: orients the line horizontally;
points are connected in order of increasing \code{x}/\code{theta} values.
\item \code{regression}: fits a linear regression line to the data.
\item \code{vertical}: orients the line vertically;
points are connected in order of increasing \code{y}/\code{r} values.
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "trees", trees)
dbGetPlot(con, "
  visualize
    age as x,
    circumference as y
  from trees
  collect by
    tree_id
  using lines
")

dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using (
     points
     layer
     regression line
  )
   scale by
     log(x),
     log(y)
")

}
