% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-ln.R
\name{ln}
\alias{ln}
\title{Natural Log}
\description{
The \code{ln} scale maps data values to a visual property
through a natural logarithm transformation.
}
\section{Function Name}{

\itemize{
\item \code{ln}
}
}

\section{Arguments}{

\itemize{
\item The name of an aesthetic to scale (required).
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)

dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using points
  scale by
    ln(x),
    ln(y)
")

}
