% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-log.R
\name{log}
\alias{log}
\title{Log}
\description{
The \code{log} scale maps data values to a visual property
through a base-10 logarithm transformation.
}
\section{Function Name}{

\itemize{
\item \code{log}
}
}

\section{Arguments}{

\itemize{
\item The name of an aesthetic to scale (required).
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)

dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using points
  scale by
    log(x),
    log(y)
")

}
