% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\docType{data}
\name{trees}
\alias{trees}
\title{Trees dataset}
\format{
A data frame with 35 rows and 3 variables:
\describe{
\item{tree_id}{Integer identifier for the tree.}
\item{age}{Age of the tree (in days) at the time of measurement.}
\item{circumference}{Trunk circumference at the time of measurement.}
}
}
\usage{
trees
}
\description{
A dataset used throughout rsgl documentation and examples
containing repeated measurements of tree growth over time.
Loading rsgl masks \link[datasets:trees]{datasets::trees}; refer to the built-in version as
\code{datasets::trees} if you need it.
}
\keyword{datasets}
