% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-violins.R
\name{violins}
\alias{violins}
\title{Violins}
\description{
Documents the aliases, aesthetics, and qualifiers for the violin geom.
}
\section{Aliases}{

\itemize{
\item \code{violin}
\item \code{violins}
}
}

\section{Aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
}
}

\section{Qualifiers}{

\itemize{
\item \code{horizontal}: orients the violins horizontally.
\item \code{vertical}: orients the violins vertically.
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    origin as x,
    miles_per_gallon as y
  from cars
  using violins
")

dbGetPlot(con, "
  visualize
    origin as y,
    miles_per_gallon as x
  from cars
  using horizontal violins
")

}
