% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-bars.R
\name{bars}
\alias{bars}
\title{Bars}
\description{
The bar geom lets you represent data using rectangular bar objects.
}
\section{Keywords}{

\itemize{
\item \code{bar}
\item \code{bars}
}
}

\section{Supported aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    bin(miles_per_gallon) as x,
    count(*) as y
  from cars
  group by
    bin(miles_per_gallon)
  using bars
")

}
