% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-boxes.R
\name{boxes}
\alias{boxes}
\title{Boxes}
\description{
The box geom lets you represent data using boxplot box objects.
}
\section{Keywords}{

\itemize{
\item \code{box}
\item \code{boxes}
}
}

\section{Supported aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    origin as x,
    miles_per_gallon as y
  from cars
  using boxes
")

}
