% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetPlot.R
\name{dbGetPlot}
\alias{dbGetPlot}
\title{Generate a plot from a SGL statement}
\usage{
dbGetPlot(con, sgl_stmt)
}
\arguments{
\item{con}{A DuckDB connection}

\item{sgl_stmt}{A SGL statement (string)}
}
\value{
The plot defined by the SGL statement (ggplot2 plot object)
}
\description{
\code{dbGetPlot} takes a DuckDB connection and a SGL statement
and returns the corresponding plot.
}
\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", mtcars)
p <- dbGetPlot(con, "
  visualize
    hp as x,
    mpg as y
  from cars
  using points
")
print(p)

}
