% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGetPlot.R
\name{dbGetPlot}
\alias{dbGetPlot}
\title{Generate a plot from a SGL statement}
\usage{
dbGetPlot(con, sgl_stmt)
}
\arguments{
\item{con}{A database connection (as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}})}

\item{sgl_stmt}{A SGL statement (string)}
}
\value{
The plot defined by the SGL statement
(a \link[ggplot2:ggplot]{ggplot2 plot})
}
\description{
\code{dbGetPlot} takes a database connection and a SGL statement
and returns the corresponding plot.
}
\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
p <- dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using points
")
print(p)

}
