% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-lines.R
\name{lines}
\alias{lines}
\title{Lines}
\description{
The line geom lets you represent data using lines.
}
\section{Keywords}{

\itemize{
\item \code{line}
\item \code{lines}
}
}

\section{Supported aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "trees", trees)
dbGetPlot(con, "
  visualize
    age as x,
    circumference as y
  from trees
  collect by
    tree_id
  using lines
")

}
