% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-points.R
\name{points}
\alias{points}
\title{Points}
\description{
Documents the aliases, aesthetics, and qualifiers for the point geom.
}
\section{Aliases}{

\itemize{
\item \code{point}
\item \code{points}
}
}

\section{Aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
\item \code{size}
}
}

\section{Qualifiers}{

\itemize{
\item \code{jittered}: adds a small amount of random noise to each point's position.
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using points
")

set.seed(0)
dbGetPlot(con, "
  visualize
    origin as x,
    miles_per_gallon as y
  from cars
  using jittered points
")

}
