% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-points.R
\name{points}
\alias{points}
\title{Points}
\description{
The point geom lets you represent data using point objects.
}
\section{Keywords}{

\itemize{
\item \code{point}
\item \code{points}
}
}

\section{Supported aesthetics}{

\itemize{
\item \code{x}
\item \code{y}
\item \code{theta}
\item \code{r}
\item \code{color}
\item \code{size}
}
}

\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "cars", cars)
dbGetPlot(con, "
  visualize
    horsepower as x,
    miles_per_gallon as y
  from cars
  using points
")

}
