% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_classifications}
\alias{type_classifications}
\title{Get SGL type classificatons for columns in a table}
\usage{
type_classifications(con, table_name)
}
\arguments{
\item{con}{A database connection (as returned by \code{\link[DBI:dbConnect]{DBI::dbConnect()}})}

\item{table_name}{The name of a table}
}
\value{
A dataframe listing the SGL type classification of each column.
}
\description{
\code{type_classifications} takes a database connection and a table name
and returns the SGL type classifications (numerical, categorical,
or temporal) of the table's columns.
}
\examples{
library(duckdb)
con <- dbConnect(duckdb())
dbWriteTable(con, "iris", iris)
type_classifications(con, "iris")

}
